+++ Code/CygSoft.SmartSession.GoalManagement/GoalTask.cs
@@ -1,9 +1,5 @@
 ﻿using CygSoft.SmartSession.GoalManagement.Infrastructure;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace CygSoft.SmartSession.GoalManagement
 {
@@ -25,7 +21,7 @@ namespace CygSoft.SmartSession.GoalManagement
             Title = title;
         }
 
-        public abstract int PercentCompleted { get; }
+        public abstract double PercentCompleted { get; }
 
         public string Title { get; private set; }
